Restructured node files, added remove group button
