+++ client/src/styles/AddGroupModal.css
@@ -1,0 +1,22 @@
+div.visible.menu.transition div.item.cell {
+    display: grid;
+    grid-template-columns: 35px 1fr;
+    align-items: center;
+}
+
+.ui.dropdown .menu>.item>img.ui.image {
+    height: 16px;
+    width: 21px;
+    border: 1px solid black;
+}
+
+.ui.dropdown .menu>.item>img.ui.image.world {
+    height: 16px;
+    width: auto;
+    padding: 1px 3.5px 1px 3.5px
+}
+
+.ui.multiple.dropdown>.ui.label {
+    display: flex;
+    align-items: center;
+}
