--- client/src/styles/AddLeagueModal.css
@@ -1,18 +1,0 @@
-img.ui.image {
-    height: 15px;
-}
-
-div.visible.menu.transition div.item.cell {
-    display: flex;
-    align-items: center;
-}
-
-.ui.label>img.ui.image {
-    height: 15px !important;
-    margin-right: .78571429rem;
-}
-
-.ui.multiple.dropdown>.ui.label {
-    display: flex;
-    align-items: center;
-}
