+++ client/src/styles/FavsGroupTable.css
@@ -3,8 +3,8 @@
 }
 
 .favGroupRow {
-    height: 50px;
-    display: flex;
+    display: grid;
+    grid-template-columns: 1fr 40px;
     align-items: center;
 }
 
@@ -41,4 +41,26 @@ a {
 img.teamLogo {
     width: 30px;
     margin-right: 10px;
+}
+
+i.icon.removeGroup {
+    color: red;
+    font-size: 1.2em;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+}
+
+/* i.icon.removeGroup:hover {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    transition: transform .7s ease-in-out;
+    transform: rotate(360deg);
+} */
+
+.ui.selectable.table tbody tr:hover i.icon.removeGroup, .ui.table tbody tr td.selectable:hover i.icon.removeGroup {
+    color: red;
+    display: contents;
 }
