+++ server/schemaModels/favTeamsModel.js
@@ -1,0 +1,10 @@
+var mongoose = require("mongoose");
+
+var favTeamsSchema = new mongoose.Schema({
+    team_id: Number,
+    name: String,
+    country: String,
+    logo: String
+});
+
+module.exports = mongoose.model("Team", favTeamsSchema);
