+++ server/views/add_league.ejs
