+++ server/views/fav_leagues.ejs
