+++ server/views/fav_teams.ejs
