+++ server/views/standings.ejs
